Ultimo commit

+++ src/trabalho/rmi/remotebase/IRemoteCalculadora.java
@@ -1,5 +1,6 @@
 package trabalho.rmi.remotebase;
 
+import java.math.BigDecimal;
 import java.rmi.Remote;
 import java.rmi.RemoteException;
 
@@ -7,8 +8,8 @@ import trabalho.rmi.cliente.Pessoa;
 
 public interface IRemoteCalculadora extends Remote {
 
-	public double verifica(Pessoa pessoa) throws RemoteException;
-	public void depositar(double valor, Pessoa pessoa) throws RemoteException;
-	public String sacar(double valor, Pessoa pessoa) throws RemoteException;
+	public BigDecimal verifica(Pessoa pessoa) throws RemoteException;
+	public void depositar(BigDecimal valor, Pessoa pessoa) throws RemoteException;
+	public String sacar(BigDecimal valor, Pessoa pessoa) throws RemoteException;
 
 }
